Teleport Code Instructions
Stand where you want to teleport. In the dolphin memory engine, look at the address 80187740 and replace the zeros in line 2 of this code with the first 8 digits in the top left. Then look at the address 80187744 and do the same thing with zeros in line 3 of this code. then look at the address 80187748 and do the same thing with the zeros in line 4.

To teleport out of bounds, set the zeros in line 3 to C7000000 .